+++ src/pages/SignInUp/Modal/Component-SignFoPw/Modal_FoPw.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/common.scss';
+@import '../../../../styles/common.scss';
 
 .Modal-signIn-overlay{
     position: fixed;
