+++ src/pages/SignInUp/Modal/Component-SignIn/Modal_SignIn.scss
@@ -1,0 +1,91 @@
+@import '../../../../styles/common.scss';
+
+        .Component-SignIn{
+          .closeBtn{
+            // overflow: hidden;
+            border:0;
+            outline:0;
+            background-color: #e7e7e7;
+            float:right;
+          }
+          
+          .mainTitle{
+            padding-bottom: 30px;
+
+            .title{
+              font-size: 16pt;
+              color: #333;
+              font-size: 33px;
+              padding: 40px 0 22px;
+              font-family: $mainFont;
+            }
+            .subTitle{
+              font-size: 19px;
+              color:#666;
+            }
+          }
+          
+          .content{
+            margin-bottom: 35px;
+            padding-top: 35px;
+
+            .TextInput{
+              width: 100%;
+              border:none;
+              background-color: #e7e7e7;
+              font-size:25px;
+              border-color:rgba(51, 51, 51, .2);
+              border-style: solid;
+              border-width: 0 0 1px 0;
+              padding-bottom:3px;     
+            }
+            input::placeholder {
+              color: rgb(155, 155, 155);
+              font-family: $mainFont;
+            }
+            .TextInput:focus{
+              border-color:rgba(102, 102, 102, 0.822);
+              border-style: solid;
+              border-width: 0 0 1px 0;
+              transition: all .2s ease-in-out;
+          }
+            input:focus::-webkit-input-placeholder, 
+            textarea:focus::-webkit-input-placeholder { 
+            /* WebKit browsers */ color:transparent; 
+            }
+          }
+      
+          .button-wrap{
+            margin: 0;
+            margin-top: 8px;
+            button{
+              width: 100%;
+              height:60px;
+              min-width: 210px;
+              background-color: #d6586b;
+              font-size: 13pt;
+              color: white;
+              border: 0;
+              // cursor: pointer;
+              
+              // &:active{
+              //   background-color: #7e49c8;
+              // }
+            }
+            .BtnColorChange{
+              width: 100%;
+              height:60px;
+              min-width: 210px;
+              background-color: $mainRed;
+              font-size: 13pt;
+              color: white;
+              border: 0;
+              transition: all .2s ease-in-out;
+              &:hover{
+                background-color: #ff1b3d;
+                transition: all .2s ease-in-out;
+              }
+            }
+          }
+        }
+
