+++ src/pages/SignInUp/Modal/Component-SignUp/Modal_SignUp.scss
@@ -1,86 +1,21 @@
-@import '../../../styles/common.scss';
+@import '../../../../styles/common.scss';
 
 
-    // .Modal-inner{
-    //     width:100%;
-        
-    //     padding-top:50px;
-    //     display:block;
+   
         
     
-        .Modal-signUp-overlay{
-            position: fixed;
-            top: 0;
-            left: 0;
-            bottom: 0;
-            right: 0;
-            background-color: rgba(88, 88, 88, 0.16);
-
-            &.Modal-anim-enter {
-                opacity: 0.00;
-                transition: all 0.2s;
-            }
-            &.Modal-anim-enter.Modal-anim-enter-active {
-                opacity: 1;
-            }
-            &.Modal-anim-leave {
-                opacity: 1;
-                transition: all 0.2s;
-            }
-            &.Modal-anim-leave.Modal-anim-leave-active {
-                opacity: 0.00;
-            }
-        }
-        .loginForm-signUp{
-            position: fixed;
-            top: 50%;
-            left: 50%;
-            transform: translate(-50%, -50%);
-            width:550px;
-            // border-radius: 10px;
-            background-color: #fffef2;
-            padding:50px;
-            box-shadow: 0px 3px 6px rgba(0,0,0,0.16);
-            
-            
-
-            &.Modal-anim-enter {
-                opacity: 0.00;
-                transform: translate(-50%, -50%) scale(0.7);
-                transition: all 0.2s;
-            }
-            &.Modal-anim-enter.Modal-anim-enter-active {
-                opacity: 1;
-                transform: translate(-50%, -50%) ;
-            }
-            &.Modal-anim-leave {
-                opacity: 1;
-                transform: translate(-50%, -50%) scale(1);
-                transition: all 0.2s;
-            }  
-            &.Modal-anim-leave.Modal-anim-leave-active {
-                opacity: 0.00;
-                transform: translate(-50%, -50%) scale(0.7);
-            }
-
-            // & > * {
-            //   &:first-child{
-            //     margin-top: 16px;
-            //   }
-            //   margin-left: 16px;
-            //   margin-right: 16px;
-            // }
+       .Component-SignUp{
 
             .gotobackBtn{
                 border:0;
                 outline:0;
-                background-color: #fffef2;
+                background-color: #e7e7e7;
             }
             .closeBtn-signUp{
                 // overflow: hidden;
                 border:0;
                 outline:0;
-                background-color: #fffef2;
+                background-color: #e7e7e7;
                 float:right;
             }
 
@@ -107,9 +42,9 @@
                 .TextInput{
                     width: 100%;
                     border:none;
-                    background-color: #fffef2;
+                    background-color: #e7e7e7;
                     font-size:25px;
-                    border-color:rgba(146, 146, 146, 0.397);
+                    border-color:rgba(51, 51, 51, .2);
                     border-style: solid;
                     border-width: 0 0 1px 0;
                     padding-bottom:3px     
@@ -138,9 +73,9 @@
                     .TextInput{
                         width: 100%;
                         border:none;
-                        background-color: #fffef2;
+                        background-color: #e7e7e7;
                         font-size:25px;
-                        border-color:rgba(146, 146, 146, 0.397);
+                        border-color:rgba(51, 51, 51, .2);
                         border-style: solid;
                         border-width: 0 0 1px 0;
                         padding-bottom:3px     
@@ -166,9 +101,9 @@
                     .TextInput{
                         width: 100%;
                         border:none;
-                        background-color: #fffef2;
+                        background-color: #e7e7e7;
                         font-size:25px;
-                        border-color:rgba(146, 146, 146, 0.397);
+                        border-color:rgba(51, 51, 51, .2);
                         border-style: solid;
                         border-width: 0 0 1px 0;
                         padding-bottom:3px     
@@ -205,9 +140,9 @@
                     .TextInput{
                         width: 70%;
                         border:none;
-                        background-color: #fffef2;
+                        background-color: #e7e7e7;
                         font-size:25px;
-                        border-color:rgba(146, 146, 146, 0.397);
+                        border-color:rgba(51, 51, 51, .2);
                         border-style: solid;
                         border-width: 0 0 1px 0;
                         padding-bottom:3px  
@@ -236,7 +171,7 @@
                 width: 100%;
                 height:60px;
                 min-width: 210px;
-                background-color: #999;
+                background-color: #d6586b;
                 font-size: 13pt;
                 color: white;
                 border: 0;
@@ -253,13 +188,13 @@
                 width: 100%;
                 height:60px;
                 min-width: 210px;
-                background-color: #333;
+                background-color: $mainRed;
                 font-size: 13pt;
                 color: white;
                 border: 0;
                 transition: all .2s ease-in-out;
                 &:hover{
-                    background-color: rgb(22, 22, 22);
+                    background-color: #ff1b3d;
                     transition: all .2s ease-in-out;
                 }
                 }
@@ -272,14 +207,12 @@
               .havaAccountBtn{
                   border:0;
                   outline: 0;
-                  background-color: #fffef2;
+                  background-color: #e7e7e7;
                   color:rgb(131, 130, 130);
                   font-family: $mainFont;
                   font-weight: 500;
               }
             }
-
         }
-            
-    //}
-    
+        
+
