--- src/pages/SignInUp/Modal/Modal_PwCo.scss
@@ -1,194 +1,0 @@
-@import '../../../styles/common.scss';
-
-.Modal-PwCo-overlay{
-    position: fixed;
-    top: 0;
-    left: 0;
-    bottom: 0;
-    right: 0;
-    background-color: rgba(88, 88, 88, 0.16);
-
-    &.Modal-anim-enter {
-        opacity: 0.00;
-        transition: all 0.2s;
-      }
-      &.Modal-anim-enter.Modal-anim-enter-active {
-        opacity: 1;
-      }
-      &.Modal-anim-leave {
-        opacity: 1;
-        transition: all 0.2s;
-      }
-      &.Modal-anim-leave.Modal-anim-leave-active {
-        opacity: 0.00;
-      }
-  }
-  .loginForm-PwCo{
-      position: fixed;
-      top: 50%;
-      left: 50%;
-      transform: translate(-50%, -50%);
-      width:550px;
-      // border-radius: 10px;
-      background-color: #fffef2;
-      padding:50px;
-      box-shadow: 0px 3px 6px rgba(0,0,0,0.16);
-      
-
-      &.Modal-anim-enter {
-        opacity: 0.00;
-        transform: translate(-50%, -50%) scale(0.7);
-        transition: all 0.2s;
-      }
-      &.Modal-anim-enter.Modal-anim-enter-active {
-        opacity: 1;
-        transform: translate(-50%, -50%) ;
-      }
-      &.Modal-anim-leave {
-        opacity: 1;
-        transform: translate(-50%, -50%) scale(1);
-        transition: all 0.2s;
-      }  
-      &.Modal-anim-leave.Modal-anim-leave-active {
-        opacity: 0.00;
-        transform: translate(-50%, -50%) scale(0.7);
-      }
-
-      // & > * {
-      //   &:first-child{
-      //     margin-top: 16px;
-      //   }
-      //   margin-left: 16px;
-      //   margin-right: 16px;
-      // }
-
-      
-      .mainTitle{
-        padding-bottom: 30px;
-
-        .title{
-          font-size: 16pt;
-          color: #333;
-          font-size: 33px;
-          padding: 40px 0 22px;
-          font-family: $mainFont;
-        }
-        .subTitle{
-          font-size: 19px;
-          color:#666;
-        }
-      }
-      
-      .content{
-        margin-bottom: 35px;
-        padding-top: 35px;
-
-        .TextInput{
-          width: 100%;
-          border:none;
-          background-color: #fffef2;
-          font-size:25px;
-          border-color:rgba(51, 51, 51, .2);
-          border-style: solid;
-          border-width: 0 0 1px 0;
-          padding-bottom:3px     
-        }
-        input::placeholder {
-            color: rgb(155, 155, 155);
-            font-family: $mainFont;
-          }
-        .TextInput:focus{
-            border-color:rgba(102, 102, 102, 0.822);
-            border-style: solid;
-            border-width: 0 0 1px 0;
-            transition: all .2s ease-in-out;
-        }
-        input:focus::-webkit-input-placeholder, 
-        textarea:focus::-webkit-input-placeholder { 
-        /* WebKit browsers */ color:transparent; 
-         }
-      }
-
-      .addContent{
-        margin-bottom: 35px;
-        padding-top: 35px;
-
-        .TextInput{
-          width: 100%;
-          border:none;
-          background-color: #fffef2;
-          font-size:25px;
-          border-color:rgba(51, 51, 51, .2);
-          border-style: solid;
-          border-width: 0 0 1px 0;
-          padding-bottom:3px     
-        }
-        input::placeholder {
-            color: rgb(155, 155, 155);
-            font-family: $mainFont;
-          }
-        .TextInput:focus{
-            border-color:rgba(102, 102, 102, 0.822);
-            border-style: solid;
-            border-width: 0 0 1px 0;
-            transition: all .2s ease-in-out;
-        }
-        input:focus::-webkit-input-placeholder, 
-        textarea:focus::-webkit-input-placeholder { 
-        /* WebKit browsers */ color:transparent; 
-         }
-         .ForgottenPwBtn{
-            margin-top:20px;
-            margin-bottom:35px;
-            float: right;
-            color:#666;
-            font-family: $mainFont;
-            font-weight: 500;
-            border:0;
-            outline:0;
-            background-color:#fffef2 ;
-         }
-      }
-  
-      .button-wrap{
-        margin: 0;
-        margin-top: 8px;
-        button{
-          width: 100%;
-          height:60px;
-          min-width: 210px;
-          background-color: #999;
-          font-size: 13pt;
-          color: white;
-          border: 0;
-          // cursor: pointer;
-          
-          // &:active{
-          //   background-color: #7e49c8;
-          // }
-        }
-        .BtnColorChange{
-          width: 100%;
-          height:60px;
-          min-width: 210px;
-          background-color: #333;
-          font-size: 13pt;
-          color: white;
-          border: 0;
-          transition: all .2s ease-in-out;
-          &:hover{
-            background-color: rgb(22, 22, 22);
-            transition: all .2s ease-in-out;
-          }
-        }
-      }
-    }
-  .closeBtn{
-    // overflow: hidden;
-    border:0;
-    outline:0;
-    background-color: #fffef2;
-    float:right;
-  }
-
-
