--- src/pages/SignInUp/Modal/Modal_SignUp_com.scss
@@ -1,102 +1,0 @@
-@import '../../../styles/common.scss';
-
-
-
-            .addContent{   
-                .content2{
-                    margin-bottom: 35px;
-                    padding-top: 35px;
-                    .TextInput{
-                        width: 100%;
-                        border:none;
-                        background-color: #fffef2;
-                        font-size:25px;
-                        border-color:rgba(146, 146, 146, 0.397);
-                        border-style: solid;
-                        border-width: 0 0 1px 0;
-                        padding-bottom:3px     
-                    }
-                    input::placeholder {
-                        color: rgb(155, 155, 155);
-                        font-family: $mainFont;
-                      }
-                    .TextInput:focus{
-                        border-color:rgba(102, 102, 102, 0.822);
-                        border-style: solid;
-                        border-width: 0 0 1px 0;
-                        transition: all .2s ease-in-out;
-                    }
-                    input:focus::-webkit-input-placeholder, 
-                    textarea:focus::-webkit-input-placeholder { 
-                    /* WebKit browsers */ color:transparent; 
-                    }
-                }
-                .content3{
-                    margin-bottom: 35px;
-                    padding-top: 35px;
-                    .TextInput{
-                        width: 100%;
-                        border:none;
-                        background-color: #fffef2;
-                        font-size:25px;
-                        border-color:rgba(146, 146, 146, 0.397);
-                        border-style: solid;
-                        border-width: 0 0 1px 0;
-                        padding-bottom:3px     
-                    }
-                    input::placeholder {
-                        color: rgb(155, 155, 155);
-                        font-family: $mainFont;
-                      }
-
-                    .TextInput:focus{
-                        border-color:rgba(102, 102, 102, 0.822);
-                        border-style: solid;
-                        border-width: 0 0 1px 0;
-                        transition: all .2s ease-in-out;
-                    }
-                    input:focus::-webkit-input-placeholder, 
-                    textarea:focus::-webkit-input-placeholder { 
-                    /* WebKit browsers */ color:transparent; 
-                     }
-                }
-
-                .content4{
-                    display:flex;
-
-                    .content4-1{
-                        margin-bottom: 35px;
-                        padding-top: 35px;
-                    }
-                    .content4-2{
-                        margin-bottom: 35px;
-                        padding-top: 35px;
-                    }
-
-                    .TextInput{
-                        width: 70%;
-                        border:none;
-                        background-color: #fffef2;
-                        font-size:25px;
-                        border-color:rgba(146, 146, 146, 0.397);
-                        border-style: solid;
-                        border-width: 0 0 1px 0;
-                        padding-bottom:3px  
-                    }
-                    input::placeholder {
-                        color: rgb(155, 155, 155);
-                        font-family: $mainFont;
-                      }
-                    .TextInput:focus{
-                        border-color:rgba(102, 102, 102, 0.822);
-                        border-style: solid;
-                        border-width: 0 0 1px 0;
-                        transition: all .2s ease-in-out;
-                    }
-                    input:focus::-webkit-input-placeholder, 
-                    textarea:focus::-webkit-input-placeholder { 
-                    /* WebKit browsers */ color:transparent; 
-                  }
-                }
-            }
-
