+++ src/pages/SignInUp/Modal/SignForm.scss
@@ -1,0 +1,65 @@
+@import '../../../styles/common.scss';
+.SignIn{
+    &.Modal-anim-enter {
+      opacity: 0.00;
+      transition: all 0.2s;
+    }
+    &.Modal-anim-enter.Modal-anim-enter-active {
+      opacity: 1;
+    }
+    &.Modal-anim-leave {
+      opacity: 1;
+      transition: all 0.2s;
+    }
+    &.Modal-anim-leave.Modal-anim-leave-active {
+      opacity: 0.00;
+    }
+  
+
+    .Modal-signIn-overlay{
+        position: fixed;
+        top: 0;
+        left: 0;
+        bottom: 0;
+        right: 0;
+        background-color: rgba(88, 88, 88, 0.589);
+    }
+      
+    .loginForm{
+        position: fixed;
+        top: 50%;
+        left: 50%;
+        transform: translate(-50%, -50%);
+        width:550px;
+        // border-radius: 10px;
+        background-color: #e7e7e7;
+        padding:50px;
+        box-shadow: 0px 3px 6px rgba(0,0,0,0.16);
+        
+
+        &.Modal-anim-enter {
+          opacity: 0.00;
+          transform: translate(-50%, -50%) scale(0.7);
+          transition: all 0.2s;
+        }
+        &.Modal-anim-enter.Modal-anim-enter-active {
+          opacity: 1;
+          transform: translate(-50%, -50%) ;
+        }
+        &.Modal-anim-leave {
+          opacity: 1;
+          transform: translate(-50%, -50%) scale(1);
+          transition: all 0.2s;
+        }  
+        &.Modal-anim-leave.Modal-anim-leave-active {
+          opacity: 0.00;
+          transform: translate(-50%, -50%) scale(0.7);
+        }
+
+       
+    }
+        
+}
+    
+  
+
